新增不执行删除日志设定

+++ src/com/db/log4j/DeleteLogs.java
@@ -24,51 +24,55 @@ public class DeleteLogs {
 	 * @param path
 	 *            日志文件路径
 	 * @param day
-	 *            日志保留天数
+	 *            日志保留天数 -1为不执行删除
 	 * @param logName
 	 *            保留的日志文件
 	 */
 	public static void deleteLog(String path, int day, String... logName) {
-		Date date;
-		Date dateNow;
-		Calendar cal = Calendar.getInstance();
-		Calendar calNow = Calendar.getInstance();
-		if (path != "" && path != null) {
-			File filepath = new File(path);
-			File[] files;
-			if (filepath.exists()) {
-				if (filepath.isDirectory()) {
-					files = filepath.listFiles();
-					for (int i = 0; i < files.length; i++) {
-						date = new Date(files[i].lastModified());
-						dateNow = new Date();
-						cal.setTime(date);
-						calNow.setTime(dateNow);
-						int year = cal.get(Calendar.YEAR);
-						int yearNow = calNow.get(Calendar.YEAR);
-						int dec = 0;
-						if (yearNow - year > 0) {
-							int days = new GregorianCalendar().isLeapYear(year) ? 366 : 365;
-							dec = calNow.get(Calendar.DAY_OF_YEAR) + days - cal.get(Calendar.DAY_OF_YEAR);
-						} else {
-							dec = calNow.get(Calendar.DAY_OF_YEAR) - cal.get(Calendar.DAY_OF_YEAR);
-						}
-						// 删除指定日期的文件
-						if (dec >= day) {
-							// 不允许删除log4j基础日志文件
-							if (logName != null) {
-								for (String s : logName) {
-									if (files[i].getName().equals(s)) {
-										continue;
+		if (day != -1) {
+			Date date;
+			Date dateNow;
+			Calendar cal = Calendar.getInstance();
+			Calendar calNow = Calendar.getInstance();
+			if (path != "" && path != null) {
+				File filepath = new File(path);
+				File[] files;
+				if (filepath.exists()) {
+					if (filepath.isDirectory()) {
+						files = filepath.listFiles();
+						for (int i = 0; i < files.length; i++) {
+							date = new Date(files[i].lastModified());
+							dateNow = new Date();
+							cal.setTime(date);
+							calNow.setTime(dateNow);
+							int year = cal.get(Calendar.YEAR);
+							int yearNow = calNow.get(Calendar.YEAR);
+							int dec = 0;
+							if (yearNow - year > 0) {
+								int days = new GregorianCalendar().isLeapYear(year) ? 366 : 365;
+								dec = calNow.get(Calendar.DAY_OF_YEAR) + days - cal.get(Calendar.DAY_OF_YEAR);
+							} else {
+								dec = calNow.get(Calendar.DAY_OF_YEAR) - cal.get(Calendar.DAY_OF_YEAR);
+							}
+							// 删除指定日期的文件
+							if (dec >= day) {
+								// 不允许删除log4j基础日志文件
+								if (logName != null) {
+									for (String s : logName) {
+										if (files[i].getName().equals(s)) {
+											continue;
+										}
 									}
 								}
+								files[i].delete();
+								log.info("############ 删除日志文件 " + files[i].getName() + " ############");
 							}
-							files[i].delete();
-							log.info("############ 删除日志文件 " + files[i].getName() + " ############");
 						}
 					}
 				}
 			}
+		} else {
+			log.info("不执行删除日志");
 		}
 	}
